Moved getPublicationDate to be a method of BibtexEntry (and moved corresponding tests)
